requirements.txt
# AI ニュース収集・翻訳ボット 依存関係
# pip install -r requirements.txt

# RSSフィード解析
feedparser>=6.0.0

# Google Gemini API
google-genai>=0.3.0

# 日時解析
python-dateutil>=2.8.0

# 環境変数管理
python-dotenv>=1.0.0

# LINE Messaging API
line-bot-sdk>=3.0.0

# Marketing Engine (UI)
streamlit>=1.30.0

# Reports & Media (Added during migration fix)
edge-tts
moviepy
xhtml2pdf
reportlab
Pillow
lxml
tweepy
requests>=2.31.0
beautifulsoup4>=4.12.0

# テストフレームワーク
pytest>=7.0.0
